refactor(gameplay): extract action registry from engine/main
